refactor(lanes): empty __init__ re-exports; consumers import submodules directly
Remove all re-exports from bloqade.lanes.__init__ (GeminiLogicalSimulator,
GeminiLogicalSimulatorTask, DetectorResult, Result, Metrics,
generate_logical_noise_model, generate_simple_noise_model, NoiseModelABC).
Replace with a module docstring only.

Repoint the 5 consumer sites that imported GeminiLogicalSimulator or
steane7_m2dets/steane7_m2obs from bloqade.lanes:
  - demo/cudaq_demo.py → bloqade.gemini + bloqade.gemini.steane_defaults
  - demo/msd_postselection_experiment.py → bloqade.gemini
  - demo/simulator_device_demo.{py,ipynb} → bloqade.gemini
  - python/tests/gemini/test_steane_transversal_x_flip_regression.py → bloqade.gemini

lanes→gemini import edges: 11 (all Tier-2 tolerated; docstring mention adds 1
to raw grep count giving 12, but no new structural edge was introduced).

Closes part of #801 (Goal 1). Goal 2 (full lanes->gemini decoupling) tracked separately.
