+++ python/bloqade/lanes/__init__.py
@@ -1,13 +1,5 @@
-from bloqade.gemini.device import (
-    DetectorResult as DetectorResult,
-    GeminiLogicalSimulator as GeminiLogicalSimulator,
-    GeminiLogicalSimulatorTask as GeminiLogicalSimulatorTask,
-    Result as Result,
-)
+"""bloqade.lanes — machine-agnostic movement compilation.
 
-from .metrics import Metrics as Metrics
-from .noise_model import (
-    generate_logical_noise_model as generate_logical_noise_model,
-    generate_simple_noise_model as generate_simple_noise_model,
-)
-from .rewrite.move2squin.noise import NoiseModelABC as NoiseModelABC
+Import submodules directly (e.g. ``from bloqade.lanes.transform import LogicalPipeline``).
+Gemini-machine specifics live in ``bloqade.gemini``.
+"""
