+++ python/tests/gemini/test_steane_transversal_x_flip_regression.py
@@ -19,9 +19,8 @@ import pytest
 from bloqade.tsim import Circuit as TsimCircuit
 
 from bloqade import qubit, squin
-from bloqade.gemini import logical as gemini_logical
+from bloqade.gemini import GeminiLogicalSimulator, logical as gemini_logical
 from bloqade.gemini.logical.stdlib import default_post_processing
-from bloqade.lanes import GeminiLogicalSimulator
 
 # Shot counts kept modest. Per the issue the expected P(obs[0] = 0) is
 # deterministic 100% on both paths (the post-selected subset on the
